+++ homeassistant/components/livisi/config_flow.py
@@ -18,7 +18,7 @@ class LivisiFlowHandler(config_entries.ConfigFlow, domain=DOMAIN):
 
     def __init__(self) -> None:
         """Create the configuration file."""
-        self._host: str = ""
+        self.host: str | None
         self._info: dict[str, Any] = {}
         self._device_info: dict[str, Any] = {}
         self._auth_headers: dict[str, Any] = {}
@@ -37,9 +37,10 @@ class LivisiFlowHandler(config_entries.ConfigFlow, domain=DOMAIN):
         """Handle the initial step."""
         errors: dict[str, str] = {}
         if user_input is not None:
+            ip_password: dict[str, str] = user_input
             try:
-                ip_password: dict[str, str] = user_input
-                return await self.async_step_credentials(ip_password)
+                await self._login(ip_password)
+                return await self.create_entity(ip_password)
             except livisi_errors.WrongCredentialException:
                 errors["base"] = "wrong_password"
             except livisi_errors.ShcUnreachableException:
@@ -48,25 +49,23 @@ class LivisiFlowHandler(config_entries.ConfigFlow, domain=DOMAIN):
                 errors["base"] = "wrong_ip_address"
 
         return self.async_show_form(
-            step_id="credentials", data_schema=self.data_schema, errors=errors
+            step_id="user", data_schema=self.data_schema, errors=errors
         )
 
-    async def async_step_credentials(self, user_input: dict[str, str]) -> FlowResult:
-        """Handle the login step."""
-        return await self._login(user_input)
-
-    async def _login(self, user_input: dict[str, str]) -> FlowResult:
-        """Login into Livisi Smart Home and register the controller."""
-        self._host = str(user_input.get(CONF_HOST))
+    async def _login(self, user_input: dict[str, str]) -> None:
+        """Login into Livisi Smart Home."""
+        self.host = user_input.get(CONF_HOST)
         web_session = aiohttp_client.async_get_clientsession(self.hass)
         self._aio_livisi = AioLivisi(web_session)
         livisi_connection_data = {
-            "ip_address": self._host,
+            "ip_address": self.host,
             "password": user_input.get(CONF_PASSWORD),
         }
 
         await self._aio_livisi.async_set_token(livisi_connection_data)
 
+    async def create_entity(self, user_input: dict[str, str]) -> FlowResult:
+        """Create LIVISI entity."""
         controller_info = await self._aio_livisi.async_get_controller()
 
         if (controller_data := controller_info.get("gateway")) is None:
@@ -82,7 +81,7 @@ class LivisiFlowHandler(config_entries.ConfigFlow, domain=DOMAIN):
             title=f"SHC {controller_type}",
             data={
                 **user_input,
-                CONF_HOST: self._host,
+                CONF_HOST: self.host,
                 "os_version": controller_info.get("osVersion"),
                 "config_version": controller_info.get("configVersion"),
             },
