+++ homeassistant/components/livisi/switch.py
@@ -1,4 +1,6 @@
 """Code to handle a Livisi switches."""
+from __future__ import annotations
+
 from typing import Any
 
 from homeassistant.components.switch import SwitchEntity
@@ -7,15 +9,10 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .const import (
-    DOMAIN,
-    LIVISI_DISCOVERY_NEW,
-    LIVISI_REACHABILITY_CHANGE,
-    LIVISI_STATE_CHANGE,
-    LOGGER,
-)
-from .shc import SHC
+from .const import DOMAIN, LIVISI_DISCOVERY_NEW, LOGGER
+from .coordinator import LivisiDataUpdateCoordinator
 
 
 async def async_setup_entry(
@@ -24,12 +21,12 @@ async def async_setup_entry(
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up switch device."""
-    shc = hass.data[DOMAIN][config_entry.entry_id]
+    coordinator: LivisiDataUpdateCoordinator = hass.data[DOMAIN][config_entry.entry_id]
 
     async def async_discover_device(device) -> None:
         """Add switch."""
         livisi_switch: SwitchEntity = await create_entity(
-            hass, config_entry, device, shc
+            hass, config_entry, device, coordinator
         )
         LOGGER.debug("Include device type: %s", device.get("type"))
         async_add_entities([livisi_switch])
@@ -40,18 +37,21 @@ async def async_setup_entry(
 
 
 async def create_entity(
-    hass: HomeAssistant, config_entry: ConfigEntry, device: dict[str, Any], shc: SHC
+    hass: HomeAssistant,
+    config_entry: ConfigEntry,
+    device: dict[str, Any],
+    coordinator: LivisiDataUpdateCoordinator,
 ) -> SwitchEntity:
     """Create Switch Entity."""
     config_details: dict[str, Any] = device["config"]
     capabilities: list = device["capabilities"]
-    is_on = await shc.async_get_pss_state(capability=capabilities[0])
+    is_on = await coordinator.async_get_pss_state(capability=capabilities[0])
     room_id: str = device["location"].removeprefix("/location/")
-    room_name: str = shc.rooms[room_id]
+    room_name: str = coordinator.rooms[room_id]
     livisi_switch = LivisiSwitch(
         hass,
         config_entry,
-        shc,
+        coordinator,
         unique_id=device.get("id"),
         manufacturer=device.get("manufacturer"),
         product=device.get("product"),
@@ -65,14 +65,14 @@ async def create_entity(
     return livisi_switch
 
 
-class LivisiSwitch(SwitchEntity):
+class LivisiSwitch(SwitchEntity, CoordinatorEntity):
     """Represents the Livisi Switch."""
 
     def __init__(
         self,
         hass,
-        config_entry,
-        shc,
+        config_entry: ConfigEntry,
+        coordinator: LivisiDataUpdateCoordinator,
         unique_id,
         manufacturer,
         product,
@@ -87,7 +87,6 @@ class LivisiSwitch(SwitchEntity):
         """Initialize the Livisi Switch."""
         self.hass: HomeAssistant = hass
         self.config_entry: ConfigEntry = config_entry
-        self._shc: SHC = shc
         self._attr_unique_id = unique_id
         self._manufacturer = manufacturer
         self._product = product
@@ -99,7 +98,7 @@ class LivisiSwitch(SwitchEntity):
         self._attr_is_on = is_on
         self._room = room
         self._version = version
-        self.aio_livisi = shc.aiolivisi
+        self.aio_livisi = coordinator.aiolivisi
         if is_on is None:
             self._attr_is_available = False
         else:
@@ -113,15 +112,7 @@ class LivisiSwitch(SwitchEntity):
             sw_version=self._version,
             via_device=(DOMAIN, str(self._attr_unique_id)),
         )
-
-        config_entry.async_on_unload(
-            async_dispatcher_connect(hass, LIVISI_STATE_CHANGE, self.update_states)
-        )
-        config_entry.async_on_unload(
-            async_dispatcher_connect(
-                hass, LIVISI_REACHABILITY_CHANGE, self.update_reachability
-            )
-        )
+        super().__init__(coordinator)
 
     @property
     def should_poll(self) -> bool:
@@ -129,16 +120,16 @@ class LivisiSwitch(SwitchEntity):
         return False
 
     @property
-    def is_on(self):
+    def is_on(self) -> bool | None:
         """Return the device state."""
         return self._attr_is_on
 
     @property
-    def available(self):
+    def available(self) -> bool:
         """Return if switch is available."""
         return self._attr_is_available
 
-    async def async_turn_on(self, **kwargs):
+    async def async_turn_on(self, **kwargs: Any) -> None:
         """Turn the entity on."""
         if self.is_on is True:
             return
@@ -148,7 +139,7 @@ class LivisiSwitch(SwitchEntity):
         if response is None:
             self._attr_is_available = False
 
-    async def async_turn_off(self, **kwargs):
+    async def async_turn_off(self, **kwargs: Any) -> None:
         """Turn the entity off."""
         if self.is_on is False:
             return
@@ -158,9 +149,17 @@ class LivisiSwitch(SwitchEntity):
         if response is None:
             self._attr_is_available = False
 
+    async def async_added_to_hass(self) -> None:
+        """Register callbacks."""
+        self.coordinator.async_add_listener(self.update_states)
+        self.coordinator.async_add_listener(self.update_reachability)
+
     @callback
-    def update_states(self, device_id_state: dict) -> None:
+    def update_states(self) -> None:
         """Update the states of the switch device."""
+        device_id_state = self.coordinator.data
+        if device_id_state is None:
+            return
         if device_id_state.get("id") != self._capability_id:
             return
         on_state = device_id_state.get("state")
@@ -173,8 +172,11 @@ class LivisiSwitch(SwitchEntity):
         self.async_write_ha_state()
 
     @callback
-    def update_reachability(self, device_id_reachability: dict) -> None:
+    def update_reachability(self) -> None:
         """Update the reachability of the switch device."""
+        device_id_reachability = self.coordinator.data
+        if device_id_reachability is None:
+            return
         if device_id_reachability.get("id") != self.unique_id:
             return
         if device_id_reachability.get("is_reachable") is False:
